+++ Entidades/CavaleiroCorrompido.h
@@ -7,7 +7,7 @@ class CavaleiroCorrompido : public Inimigo {
 public:
     CavaleiroCorrompido();
 
-    int ataque() const override;
+    float ataque() const override;
     int Dano_ataque(Entidade& vitima) const override;
     void receber_dano(int dano) override;
