+++ Entidades/Guerreiro.h
@@ -14,7 +14,7 @@ class Guerreiro : public Entidade{
         int get_vidaMaxima()const{return vidaMaxima;}
         string get_armaEquipada_nome()const{return arma->get_nome_arma();}
         void equipar_arma(Arma* eqp)override{arma = eqp;}
-        int ataque()const;
+        float ataque()const;
         int Dano_ataque(Entidade&)const;
         void receber_dano(int)override;
         void buffVida(int vida)override{ vidaBase += vida;};
